requirements.txt
# Core LLM providers
openai>=1.0.0
anthropic>=0.7.0
groq>=0.4.0
google-generativeai>=0.3.0
cerebras-cloud-sdk>=0.1.0

# Local LLM support
requests>=2.31.0

# Existing dependencies
chromadb>=0.4.0
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.1.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
fuzzywuzzy>=0.18.0
langdetect>=1.0.9

